Fix lint error
